# Improve resize_bicubic performance by reorganizing loops