--- a/tensorflow/core/kernels/resize_bicubic_op_test.cc
+++ b/tensorflow/core/kernels/resize_bicubic_op_test.cc
@@ -287,13 +287,13 @@ BM_ResizeBicubicDev(32, 512, 3);
 BM_ResizeBicubicDev(32, 1024, 3);
 
 #define BM_ResizeBicubicExpand(BATCH, SIZE, CHANNELS)                          \
-  static void BM_ResizeBicubicExpand####BATCH####SIZE####CHANNELS(int iters) { \
+  static void BM_ResizeBicubicExpand##_##BATCH##_##SIZE##_##CHANNELS(int iters) { \
     testing::ItemsProcessed(static_cast<int64>(iters) * BATCH * SIZE * SIZE *  \
                             CHANNELS * 8 * 8);                                 \
     test::Benchmark("cpu", ResizeBicubic(BATCH, SIZE, CHANNELS, 8, 8))         \
         .Run(iters);                                                           \
   }                                                                            \
-  BENCHMARK(BM_ResizeBicubicExpand####BATCH####SIZE####CHANNELS);
+  BENCHMARK(BM_ResizeBicubicExpand##_##BATCH##_##SIZE##_##CHANNELS);
 
 BM_ResizeBicubicExpand(12, 48, 1);
 BM_ResizeBicubicExpand(12, 48, 3);