@yongtang 2017-10-20
Thanks @cwhipkey for the suggestion. The PR has been updated. The benchmark is the following:
```
BEFORE PR:

INFO: Running command line: bazel-bin/tensorflow/core/kernels/resize_bicubic_op_test '--benchmarks=all' '--benchmark_min_time=20'
Running main() from test_main.cc
Benchmark                      Time(ns) Iterations
--------------------------------------------------
BM_ResizeBicubic_8_32_3           45170      15505       544.1M items/s
BM_ResizeBicubic_8_128_3         297522       2338       1321.6M items/s
BM_ResizeBicubic_8_512_3        5093420        100       1235.2M items/s
BM_ResizeBicubic_8_1024_3      20767140        100       1211.8M items/s
BM_ResizeBicubic_16_32_3          63796      10000       770.4M items/s
BM_ResizeBicubic_16_128_3        573589       1000       1371.1M items/s
BM_ResizeBicubic_16_512_3      10392390        100       1210.8M items/s
BM_ResizeBicubic_16_1024_3     48047780        100       1047.5M items/s
BM_ResizeBicubic_32_32_3         100327       6961       979.8M items/s
BM_ResizeBicubic_32_128_3       1130870        568       1390.8M items/s
BM_ResizeBicubic_32_512_3      20679990        100       1216.9M items/s
BM_ResizeBicubic_32_1024_3     97105890        100       1036.6M items/s
BM_ResizeBicubicExpand12481     4472045        156       395.7M items/s
BM_ResizeBicubicExpand12483     8055270        100       659.0M items/s
BM_ResizeBicubicExpand124840  308084570        100       229.7M items/s

AFTER PR:

INFO: Running command line: bazel-bin/tensorflow/core/kernels/resize_bicubic_op_test '--benchmarks=all' '--benchmark_min_time=20'
Running main() from test_main.cc
Benchmark                      Time(ns) Iterations
--------------------------------------------------
BM_ResizeBicubic_8_32_3           45280      15582       542.8M items/s
BM_ResizeBicubic_8_128_3         295346       2353       1331.4M items/s
BM_ResizeBicubic_8_512_3        5118330        100       1229.2M items/s
BM_ResizeBicubic_8_1024_3      20398880        100       1233.7M items/s
BM_ResizeBicubic_16_32_3          62595      10000       785.2M items/s
BM_ResizeBicubic_16_128_3        570425       1000       1378.7M items/s
BM_ResizeBicubic_16_512_3      10208090        100       1232.6M items/s
BM_ResizeBicubic_16_1024_3     48496160        100       1037.8M items/s
BM_ResizeBicubic_32_32_3         100469       6977       978.5M items/s
BM_ResizeBicubic_32_128_3       1121702        564       1402.2M items/s
BM_ResizeBicubic_32_512_3      20287840        100       1240.4M items/s
BM_ResizeBicubic_32_1024_3     96245560        100       1045.9M items/s
BM_ResizeBicubicExpand12481     6819970        100       259.5M items/s
BM_ResizeBicubicExpand12483     7976640        100       665.5M items/s
BM_ResizeBicubicExpand124840  212033740        100       333.8M items/s
```