@yongtang 2017-10-20
The improvement in large `num_channels` (`= 40`)  is quite visible (`229.7M items/s` -> `333.8M items/s`)

On the other hand, for `num_channels = 1` there is a performance drop (`395.7M items/s` -> `259.5M items/s`). This is more or less expected as we have a for loop that is doing nothing.

Depending on how frequent the `num_channels = 1`, it might worth the effort to treat `num_channels = 1` as a special case as well. Let me know if this is desired and I could update the PR accordingly.