@yongtang 2017-10-20
Thanks @cwhipkey. The PR has been updated. I think it probably is OK to merge this PR for now. Much appreciated for the kindly help.