@vrv 2017-10-20
@tensorflow-jenkins test this please